moved loading of initial pcds into the sync logic
